+++ src/misc/mod.rs
@@ -1,11 +1,13 @@
 mod errors;
 mod field_inputs;
+mod iterators;
 mod mut_rc_ref_cell;
 mod parse_traits;
 mod string_conversion;
 
 pub(crate) use errors::*;
 pub(crate) use field_inputs::*;
+pub(crate) use iterators::*;
 pub(crate) use mut_rc_ref_cell::*;
 pub(crate) use parse_traits::*;
 pub(crate) use string_conversion::*;